Revert "フォルダへの格納"
This reverts commit 6dbba7036da24626ec38d773c991fccf2ccaa845.
