--- takada/hash/OnemoreException.java
@@ -1,14 +1,0 @@
-package hash;
-
-public class OnemoreException extends Exception {
-
-	OnemoreException(){
-		super("メインメニューに戻ります");
-	}
-
-	OnemoreException(String err){
-		super(err);
-	}
-
-}
-
